Move some files to lib-preference-pages
