+++ libraries/lib-preference-pages/PrefsPanel.cpp
@@ -77,7 +77,7 @@ PrefsPanel::Factories
       PathStart,
       {
          {wxT(""),
-       wxT("Device,Playback,Recording,Quality,GUI,Tracks,Directories,Warnings,Effects,KeyConfig,Mouse,Module,Application")
+       wxT("Device,Playback,Recording,Quality,GUI,Tracks,ImportExport,Directories,Warnings,Effects,KeyConfig,Mouse")
          },
          {wxT("/Tracks"), wxT("TracksBehaviors,Spectrum")},
       }
