+++ modules/import-export/mod-mp3/MP3Prefs.cpp
@@ -12,7 +12,7 @@
 #include "ExportMP3.h"
 #include "Internat.h"
 #include "ShuttleGui.h"
-#include "prefs/LibraryPrefs.h"
+#include "LibraryPrefs.h"
 #include "HelpSystem.h"
 #include "ReadOnlyText.h"
