+++ src/CMakeLists.txt
@@ -606,8 +606,6 @@ list( APPEND SOURCES
       prefs/ImportExportPrefs.h
       prefs/KeyConfigPrefs.cpp
       prefs/KeyConfigPrefs.h
-      prefs/LibraryPrefs.cpp
-      prefs/LibraryPrefs.h
       prefs/MidiIOPrefs.cpp
       prefs/MidiIOPrefs.h
       prefs/ModulePrefs.cpp
@@ -616,8 +614,6 @@ list( APPEND SOURCES
       prefs/PlaybackPrefs.h
       prefs/PrefsDialog.cpp
       prefs/PrefsDialog.h
-      prefs/PrefsPanel.cpp
-      prefs/PrefsPanel.h
       prefs/QualityPrefs.cpp
       prefs/QualityPrefs.h
       prefs/RecordingPrefs.cpp
@@ -1313,6 +1309,7 @@ set( AUDACITY_LIBRARIES
    lib-viewport-interface
    lib-wave-track-paint-interface
    lib-music-information-retrieval-interface
+   lib-preference-pages-interface
 )
 
 if (USE_VST)
